[FIX] moved all axios calls originating from editcomponent to dedicated aios service
